Fix bug in indexer unit tests.
When creating a new listing contract proxy, we need to have it created with a Listing ABI, not a ListingsRegistry. In this case, I’m defaulting to the UnitListing. I’m not adding configuration options here to choose between unit and fractional listings, since I think we are going back to a single listing contract for both types.